+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-04-05
+
+### Fixed
+
+- Fixed error when deleting a build from the profile/created-builds page.
+
 ## 2024-04-04
 
 ### Fixed